Action for the angry and total added

+++ pro/model/Flowbit.py
@@ -8,7 +8,7 @@ class FlowbitSchema(BaseModel):
     source_type:Annotated[Literal['json','email','pdf'],Field(description='This will tell about the source of the request')]
     intent:Annotated[Literal['RFQ','Invoice','Complaint'],Field(description="This will tell the intent of the request")]
     tone:Annotated[Literal['angry','extreamly angry','nice'],Field(description="The tone in which the data is written")]
-    total:Annotated[Optional[float],Field(description="give the total amout of the qoutation or invoice")]
+    total:Annotated[Optional[int],Field(description="give the total amout of the qoutation or invoice")]
     # policy:Annotated[Optional[Literal["GDPR","FDA"]],Field(description="if the policy mentions ")] 
     customer:CutomerSchema
     request:RequestSchema
